Create basic GameManager
This starts the creation of the GameManager. It will take care of setting up and handle all the phases of the game while in the respective scene.

For now just instantiating and setting up the game state machine. The states created are just placeholder and maybe change in the future.