Implement cards deal
This implements the deal of cards based on provided layout.

Did a refactor to level layout scriptable object to be more clear and change type for column and rows. Also added a space vector2 to define the world distance between cards.

The cards are being positioned centered inside a cards container.