Create UIGameManager
This implements the manager that handles the user interfaces.

Created a prefabs collection scriptable object with a list of prefabs, that allows the manager to get the prefab by its type and then instantiate.

Also added a UIBasePanel to be used as base class for all panels. Just contains a show and hide method, and then helps with the generics.